requirements.txt
psycopg2==2.9.10
djangorestframework==3.15.2
Pillow==11.1.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0